Revert "LogRecordType"
This reverts commit 78ac262b7021cef5e25db61b041a560e04cf76a2.

Revert "LayoutType"

This reverts commit c7a4ea44806217ae77b10c4b940adccd447bdafc.

Revert "fix LayoutType and ConstraintType"

This reverts commit e99c12e92dc50ff15f96ae4dc92151880cef085a.

Revert "Fix LoggingType logger"

This reverts commit 4d68c66377a78d4c13a95dbffa2a68b73fdcb431.

Move changes to other PR

+++ src/main/logger/logger_configuration.cpp
@@ -60,7 +60,7 @@ static struct option opts[] = {
 
 static void ValidateLoggingType(const configuration& state) {
   if (state.logging_type <= LoggingType::INVALID) {
-    LOG_ERROR("Invalid logging_type :: %d", static_cast<int>(state.logging_type));
+    LOG_ERROR("Invalid logging_type :: %d", state.logging_type);
     exit(EXIT_FAILURE);
   }
